Pause menu
So this push includes the change to the dialogue manager that I asked about in discord (advancing dialogue now requires that you hit N key), but the pause menu works on the bright side? If you want, you can delete the DialogueManager.cs change, or just send it back, but I really need to push this while my Unity/GitHub Desktop are being cooperative.