Select Order to Make
Player can now hit the orders button on screen and see what orders they can make. Orders display the customer they are for, the recipe of the order, and the reward for completion. Player can hit the make order button to make any of the orders on the bar,  or they can click outside of the tab to hide the tab.